Allow to invite members to an organizations, and to quit it.

+++ app/controllers/organizations_controller.ts
@@ -8,7 +8,21 @@ import type { HttpContext } from '@adonisjs/core/http'
 import mail from '@adonisjs/mail/services/main'
 
 export default class OrganizationsController {
-  public async index({ inertia }: HttpContext) {
+  public async index({ auth, inertia, response }: HttpContext) {
+    /**
+     * If the current user is related to any organization,
+     * then redirect to the first organization.
+     */
+    await auth.user!.load('organizations')
+    if (auth.user!.organizations.length) {
+      const organization = auth.user!.organizations[0]
+      return response.redirect().toPath(`/organizations/${organization.slug}/projects`)
+    }
+
+    /**
+     * If the user is not related to any organization,
+     * then render a page allowing to create a new organization.
+     */
     return inertia.render('organizations/index')
   }
 
@@ -44,6 +58,9 @@ export default class OrganizationsController {
     organization: Organization,
     organizationMember: OrganizationMember
   ) {
+    if (organizationMember.role !== 'owner') {
+      return response.unauthorized()
+    }
     const isOwner = organizationMember.role === 'owner'
     if (!isOwner) {
       return response.unauthorized()
@@ -55,7 +72,14 @@ export default class OrganizationsController {
   }
 
   @bindOrganization
-  public async destroy({ response }: HttpContext, organization: Organization) {
+  public async destroy(
+    { response }: HttpContext,
+    organization: Organization,
+    organizationMember: OrganizationMember
+  ) {
+    if (organizationMember.role !== 'owner') {
+      return response.unauthorized()
+    }
     await organization.delete()
     return response.redirect().toPath('/organizations')
   }
@@ -66,6 +90,9 @@ export default class OrganizationsController {
     _organization: Organization,
     organizationMember: OrganizationMember
   ) {
+    if (organizationMember.role !== 'member') {
+      return response.unauthorized()
+    }
     await organizationMember.delete()
     return response.redirect().toPath('/organizations')
   }
@@ -76,13 +103,12 @@ export default class OrganizationsController {
     organization: Organization,
     organizationMember: OrganizationMember
   ) {
-    if (organizationMember.role === 'owner') {
+    if (organizationMember.role !== 'owner') {
       return response.unauthorized()
     }
     const email = request.input('email')
     await mail.send(new InviteMemberNotification(organization, email))
-
-    return response.redirect().back()
+    return response.redirect().toPath(`/organizations/${organization.slug}/edit`)
   }
 
   public async join({ auth, request, response }: HttpContext) {
@@ -90,7 +116,7 @@ export default class OrganizationsController {
       return response.redirect().toPath('/auth/sign_up')
     }
     const organization = await Organization.findByOrFail('slug', request.param('organizationSlug'))
-    await organization.related('members').create({
+    await organization.related('members').firstOrCreate({
       userId: auth.user!.id,
       role: 'member',
     })
